Introduce the orchestrator and start the simulation from it
The orchestrator sends a message to the simulation_requests topic which
contains all the requests for all the engines. The engine filters its
own payload and runs the simulation.

Introduce the concept of engine_id to identify the engine. (This is
needed only when running with multiple engines).

Warning: Running multiple engines using the UI is not yet supported