Bug fix - fix overlap between areas
Area boundaries were previously being "shared". This is incorrect,
because each point is a cell, and there is no concept of a boundary.

The SVG zoom is misleading because it shows gaps in the homes/offices.
These gaps aren't visible in the PNG even when zooming 4000x. That
means the SVG is probably zooming each polygon at a different scale.

Changed the default grid image to PNG instead.